Refactor core package (#476)
